+++ CHANGELOG.md
@@ -1,3 +1,24 @@
 # Changelog
 
 All Notable changes to `event-store` will be documented in this file
+
+# 2.0
+
+* Removed MutableEventStoreInterface (issue #5)
+
+# 1.2
+
+* Added support for upcasting to multiple events
+* Deprecated support of upcasters returning non array values
+
+# 1.1.1
+
+Fixed partial stream upcasting 
+
+# 1.1
+
+Added partial stream support
+
+# 1.0 
+
+First release